Client cohort report using jinja template
