--- src/agr/gbs_prism/html/KGD-per-cohort.html.jinja
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>
-Results for {{ title }} for {{ cohort }}
-</title>
-</head>
-<body>
-<h1> Results for {{ title }} for {{ cohort }}</h1>
-
-</body>
-</html>
