+++ src/agr/gbs_prism/reports.py
@@ -1,0 +1,262 @@
+import os.path
+from dataclasses import dataclass
+from jinja2 import Environment, PackageLoader, select_autoescape
+from typing import Literal, Optional
+
+
+@dataclass(kw_only=True)
+class Row:
+    name: str
+    narration: Optional[str] = None
+    cohort_targets: dict[
+        str, str
+    ]  # for inline kind this is the file content not the path
+
+
+@dataclass
+class Group:
+    name: str
+    kind: Literal["image", "link", "inline"]
+    rows: list[Row]
+
+
+@dataclass
+class Report:
+    title: str
+    cohorts: list[str]
+    groups: list[Group]
+
+
+def render_report(report: Report, out_path: str):
+    env = Environment(
+        loader=PackageLoader("agr.gbs_prism"), autoescape=select_autoescape()
+    )
+    template = env.get_template("report.html.jinja")
+    with open(out_path, "w") as out_f:
+        _ = out_f.write(
+            template.render(
+                title=report.title,
+                cohorts=report.cohorts,
+                groups=report.groups,
+            )
+        )
+
+
+def create_report(
+    title: str,
+    cohort_target_dirs: dict[str, str],
+    make_targets_relative_to: Optional[str] = None,
+) -> Report:
+    """Create report, target dir for a cohort is the one containing KGD as a subdirectory."""
+    return Report(
+        title=title,
+        cohorts=sorted(cohort_target_dirs.keys()),
+        groups=[
+            Group(
+                name="KGD (plots)",
+                kind="image",
+                rows=[
+                    Row(
+                        name=name,
+                        cohort_targets={
+                            cohort: os.path.relpath(
+                                os.path.join(cohort_target_dir, "KGD", name),
+                                make_targets_relative_to,
+                            )
+                            for (
+                                cohort,
+                                cohort_target_dir,
+                            ) in cohort_target_dirs.items()
+                        },
+                        narration=narration,
+                    )
+                    for (name, narration) in KGD_PLOTS_WITH_NARRATION
+                ],
+            )
+        ],
+    )
+
+
+KGD_PLOTS_WITH_NARRATION = [
+    (
+        "AlleleFreq.png",
+        """
+Comparisons SNP reference allele frequencies calculated 3 different ways: 1) after naively converting to genotypes, 2) based on allele counts (without any adjustment for multiple counts of the same allele), 3) as given by UNEAK (the same as method 2).
+""",
+    ),
+    (
+        "CallRate.png",
+        """
+Histogram of call rate (proportion of SNPs scored) for each sample.
+""",
+    ),
+    (
+        "Co-call-HWdgm.05.png",
+        """
+Shows the co-call distribution (see Co-call-.png) after applying the Hardy-Weinberg filter to SNPs and combining lanes (if relevant).
+""",
+    ),
+    (
+        "Co-call-.png",
+        """
+The co-call plot shows the distribution (over all possible pairs) of the proportion of SNPs called in a pair of individuals.
+Bimodality may indicate that the samples belong to 2+ genetically divergent clusters (low co-call rate for pairs from different clusters)
+""",
+    ),
+    (
+        "finplot.png",
+        """
+A finplot shows raw Hardy-Weinberg disequilibrium (PAA - pA^2) against MAF for each SNP. 
+SNPs near the upper border tend to be mainly homozygous. This is often because they have low depth (perhaps only one allele observed).
+SNPs near the lower border tend to be mainly heterozygous and may often represent duplicated regions erroneously treated as one region. 
+The Hardy-Weinberg filter, HWdgm.05 (Hardy-Weinberg disequilibium > -0.05) used in some analyses removes these heterozygous SNPs.
+""",
+    ),
+    (
+        "GcompareHWdgm.05.png",
+        """
+Same as Gcompare.png plot but using only those SNPs passing the Hardy-Weinberg filter and combining lanes (if relevant)
+""",
+    ),
+    (
+        "Gcompare.png",
+        """
+Comparisons of different methods for calculating off-diagonals of the GRM. G5=KGD method, G3=KGD with single allele per genotype, G1=unadjusted (biased towards 0).
+""",
+    ),
+    (
+        "Gdiagdepth.png",
+        """
+In theory self-relatedness estimated from GBS data should be uncorrelated with sample depth, so there should be neither negative nor positive trend
+ in this plot. A negatively sloping trend has been nick-named a slippery slope, and means something is not quite right somewhere (with either the make-up of the original sample
+ -e.g. some kind of contamination, or the enzyme digest, or size selection, or sequencing, or downstream processing). If they persist, slippery slopes need to be investigated and
+ resolved
+""",
+    ),
+    (
+        "GHWdgm.05diagdepth.png",
+        """
+This is similar to the Gdiagdepth.png plot, but 1) combining lanes (if relevant) and 2) based on a smaller set of SNPs (net of the Hardy-Weinberg filter), which eliminates some potential sources of
+ correlation between self-relatedness estimates and sample depth. So this plot should exhibit less than (or the same) correlation as that plot. 
+""",
+    ),
+    (
+        "InbCompare.png",
+        """
+Check clumpify normalisation (deduplication) by comparing four estimates of inbreeding: a) production estimate
+ b) adjust using fitted beta-binomial (bb) c) sampling one read per lane to remove any duplication effect (used to fit bb)
+ d) treat two lanes as two individuals. If the points are below the line in the first column (a .v. b,c,d), or alpha
+ is low, our normalisation may not be aggressive enough, as production inbreeding estimate is thus higher than these other
+ three methods (each of which should provide a robust estimate of inbreeding, even with un-normalised data)
+""",
+    ),
+    (
+        "Heatmap-G5HWdgm.05.png",
+        """
+Heatmap representation of the G5 GRM.
+Relatedness values are coloured from white to red (lowest to highest relatedness). Samples are ordered by a dendrogram from hierarchically clustering a distance matrix of the GRM.
+SNPs have been Hardy-Weinberg filterd and and lanes combined (if relevant)
+""",
+    ),
+    (
+        "HWdisMAFsig.png",
+        """
+SNPs plotted as in finplot.png but coloured by an approximate depth-adjusted significance value.
+""",
+    ),
+    (
+        "MAFHWdgm.05.png",
+        """
+MAF distribution of Hardy-Weinberg filtered SNPs after combining lanes (if relevant)
+""",
+    ),
+    (
+        "MAF.png",
+        """
+MAF distribution. Diverse populations tend to have a peak at low values of MAF.
+""",
+    ),
+    (
+        "PC1v2G5HWdgm.05.png",
+        """
+Principal components plot (1st 2 coordinates) of samples, based on the G5 GRM.
+SNPs have been Hardy-Weinberg filterd and and lanes combined (if relevant)
+""",
+    ),
+    (
+        "SampDepthCR.png",
+        """
+Plot of depth vs call rate for samples. This is usually a fairly tight upwardly curving line.
+""",
+    ),
+    (
+        "SampDepthHist.png",
+        """
+Distribution of sample depths. If it is not roughly unimodal, there may be contamination or a mix of different species or breeds. Also, some downstream
+ pipelines use a sample-depth cutoff to filter the data, which could be confused by a multi-modal sample-depth distribution.
+""",
+    ),
+    (
+        "SampDepth.png",
+        """
+Plot of mean vs median sample depth.
+""",
+    ),
+    (
+        "SampDepth-scored.png",
+        """
+Plot of mean sample depth for SNPs with at least one read for the sample against the mean sample depth including all SNPs.
+""",
+    ),
+    (
+        "SNPCallRate.png",
+        """
+Histogram of the proportion of samples called for a SNP. There is usually a peak at a high value and tails off to the left.
+""",
+    ),
+    (
+        "SNPDepthHist.png",
+        """
+Histogram of mean SNP depth. There are often a few SNPs with extremely high depth, while most SNPs have relatively low depth. 
+The SNPdepth.png plot has SNP depth log-transformed to show the distribution more clearly.
+""",
+    ),
+    (
+        "SNPDepth.png",
+        """
+Plot of SNP call rate against (log-transformed) SNP depth. Usually a S-shaped band. 
+High depth low call rate SNPs may indicate diverse populations (some SNPs only seen in a subset) (and/or possibly size selection variation?)
+""",
+    ),
+    (
+        "X2star-QQ.png",
+        """
+Quantile-quantile plot of approximate depth-adjusted Hardy-Weinberg test statistic for each SNP. A 1-1 line indicates the theoretical distribution holds. 
+A higher sloped straight line suggests some population structure. SNPs that plot higher than a straight line through the majority of SNPs may indicate SNPs that do not behave in a Mendelian fashion.
+""",
+    ),
+    (
+        "PlateDepth.png",
+        """
+Plot of mean sample depth by plate position. Patterns of depth variation may represent problems with sample handling.
+""",
+    ),
+    (
+        "PlateInb.png",
+        """
+Plot of estimated inbreeding by plate position. Patterns of inbreeding variation may represent problems with sample handling.
+""",
+    ),
+    (
+        "SubplateDepth.png",
+        """
+The same as PlateDepth.png but with a different colour gradient for each subplate within the main plate.
+""",
+    ),
+    (
+        "SubplateInb.png",
+        """
+The same as PlateInb.png but with a different colour gradient for each subplate within the main plate.
+""",
+    ),
+]
