--- src/agr/gbs_prism/reports/__init__.py
@@ -1,7 +1,0 @@
-# re-exports for agr.gbs_prism.reports
-
-from .kgd import make_kgd_cohort_report
-
-__all__ = [
-    "make_kgd_cohort_report",
-]
